telegram_bot/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic==2.5.2
aiogram==3.22.0
SQLAlchemy==2.0.44
asyncpg==0.30.0